requirements.txt
Flask==2.2.3
gunicorn==20.1.0
gevent==22.10.2
psycopg2-binary==2.9.5
SQLAlchemy==2.0.5
python-dotenv==1.0.0
requests==2.28.2
beautifulsoup4==4.11.2
lxml==4.9.2
Pillow==9.4.0
pytest==7.3.1
selenium==4.8.3
webdriver-manager==3.8.5
pandas==1.5.3
numpy==1.24.2
matplotlib==3.7.1
seaborn==0.12.2
Jinja2==3.1.2
Werkzeug==2.2.3
itsdangerous==2.1.2
click==8.1.3
MarkupSafe==2.1.2
urllib3==1.26.15
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
six==1.16.0
python-dateutil==2.8.2
pytz==2023.3
tzdata==2023.3
greenlet==2.0.2
typing-extensions==4.5.0
